+++ apps/web-app/src/features/swap/hooks/useSwapExecution.ts
@@ -18,7 +18,7 @@ import {
   type CowSwapLimitOrderRequest,
 } from "@/lib/helpers/evm/cowswap";
 import { getTokensForChain } from "@/lib/constants/evmConfig";
-import { getEvmTokenSymbol } from "@/lib/helpers/stellar/tokenUtils";
+import { getEvmTokenSymbol } from "@/lib/helpers/tokenUtils";
 import { useWallet } from "@/hooks/useWallet";
 import { handleSwapError } from "@/features/swap/utils/swapErrorUtils";
